requirements.txt
streamlit
azure-cognitiveservices-speech
openai>=1.0.0
